存期修改成可以设置6个月的

+++ backend/src/main/java/com/bookkeeping/dto/CreateDepositRequest.java
@@ -26,7 +26,7 @@ public class CreateDepositRequest {
     
     private BigDecimal interestRate;
     
-    private Integer term;
+    private BigDecimal term;
     
     private String note;
